logout button added for logged in users
